backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
# psycopg2-binary==2.9.7  # Problème de compilation avec Python 3.13
# Solution: utiliser psycopg (version 3) qui a des wheels précompilés pour Python 3.13
psycopg[binary]>=3.1.0  # Version moderne compatible Python 3.13, supportée par SQLAlchemy 2.0+
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy>=2.0.23  # Version compatible Python 3.13 (2.0.21 a des problèmes avec Python 3.13)
PyJWT==2.8.0
bcrypt==4.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.3
alembic==1.12.0
Mako==1.2.4
transformers
torch
huggingface-hub
sentence-transformers
accelerate
PyPDF2
python-docx
